feat: AI photo validation for solo mode, fix game mode bugs, cleanup
Solo Mode:
- Add Gemini 2.0 Flash AI photo validation via Supabase Edge Function
- New scoring: sum of AI ratings (1-10 per photo) + time bonus
- KatchIt! AI branding with animated gradient UI
- Pulsing border animation during AI validation
- Score counter animates on rating reveal

Bug Fixes:
- Fix Blind Bingo reveal using float division for even timing
- Fix Solo WIN_ROUND daily challenge requiring at least 1 capture
- Fix Daily Challenge showing raw enum names (now localized)
- Fix Daily Challenge using UTC timezone (prevents timezone exploit)
- Add periodic midnight reset for daily challenges
- Fix Solo Leaderboard deduplication (best score per player)
- Remove non-functional QuickStart difficulty UI

Cleanup:
- Remove 37 unused AppIcon PNG files (fixes Xcode warning)
- Remove dead code quickStartCategoryPreview()
- Fix HomeScreen stars/profile alignment to right

+++ GeoBingo/composeApp/src/commonMain/kotlin/pg/geobingo/one/ui/screens/game/GameScreenContent.kt
@@ -291,7 +291,7 @@ fun GameScreenContent(
                     val elapsed = totalSeconds - gameState.gameplay.timeRemainingSeconds
                     val totalCats = gameState.gameplay.selectedCategories.size
                     val revealedCount = if (isBlindBingo && totalCats > 0 && totalSeconds > 0) {
-                        (elapsed * totalCats / totalSeconds + 1).coerceIn(1, totalCats)
+                        (elapsed.toFloat() * totalCats / totalSeconds + 1).toInt().coerceIn(1, totalCats)
                     } else {
                         totalCats
                     }
